Complete rewrite for Praw 4
Everything should be a lot cleaner now, and a lot more nicely laid out
and commented.

This is now modular for different subreddits, so it can work with
several different ones at once.

Filestorage changed to JSON, which isn't ideal, but it means I can store
more different kinds of data, like the webhooks which I'm going to
implement later.

I quite like this update. I've been meaning to do this for a while.